feat: add more deployment details

+++ scripts/reduce-large-file.main.kts
@@ -9,7 +9,7 @@ import java.io.File
 require(args.size == 2) {
     "ERROR: wrong argument count. Supplied ${args.size}: $args.\nUsage: reduce-large-file.main.kts <input> <output>"
 }
-data class Entry(val name: String, val latitude: Double, val longitude: Double, val samples: List<List<Float>>)
+data class Entry(val name: String, val latitude: Double, val longitude: Double, val samples: List<List<Double>>)
 val result = mutableListOf<Entry>()
 File(args[0]).reader().use { fileReader ->
     val reader = JsonReader(fileReader)
